requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.0.3
numpy<2.0.0
scikit-learn==1.3.0
xgboost==2.0.0
tensorflow==2.13.0
supabase==2.0.0
google-generativeai==0.3.1
chromadb==0.4.18
python-dotenv==1.0.0
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
